[oes-vla-si] ice_evp_omp: bug-fix

+++ src/sea_ice/dynamics_fem/mo_ice_elements.f90
@@ -16,9 +16,9 @@ module mo_ice_elements
   PUBLIC
   save
   integer                                      :: elem2D
-  integer(KIND=4), allocatable, dimension(:,:) :: elem2D_nodes
-  integer(KIND=4), allocatable, dimension(:,:) :: nod2D_elems
-  integer(KIND=4), allocatable, dimension(:,:) :: elem2D_nghbrs
+  integer, allocatable, dimension(:,:) :: elem2D_nodes
+  integer, allocatable, dimension(:,:) :: nod2D_elems
+  integer, allocatable, dimension(:,:) :: elem2D_nghbrs
   REAL(wp)                                 :: Vol2D
   REAL(wp), allocatable, dimension(:,:)    :: derivative_stdbf
   REAL(wp), allocatable, dimension(:,:)    :: bafux, bafuy
@@ -27,7 +27,7 @@ module mo_ice_elements
 
   ! vladimir: indices for elements/nodes with actual ice
   ! allocated/deallocated each time-step
-  integer                                      :: si_elem2D, si_nod2D
-  integer(KIND=4), allocatable, dimension(:) :: si_idx_elem, si_idx_nodes
+  integer                                    :: si_elem2D, si_nod2D
+  integer, allocatable, dimension(:)         :: si_idx_elem, si_idx_nodes
 
 end module mo_ice_elements
